Set Evaluator and Cards Macro
Evaluates sets via rule sets and specification parameters.

Macro CardsLib::Cards[] takes a list of strings and generates
instances of cards for each string.
